Fix API tests
